do not use #ifdef but crazy include instead
