+++ src/cleanup-funcs_test.c
@@ -1,0 +1,2 @@
+#include "cleanup-funcs.h"
+#include "cleanup-funcs.c"
