+++ src/main.c
@@ -14,17 +14,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-#ifdef _UNIT_TESTING
-#include "unit-tests.h"
-#else
 #include "sc-main.h"
-#endif
 
 int main(int argc, char **argv)
 {
-#ifdef _UNIT_TESTING
-	return sc_run_unit_tests(&argc, &argv);
-#else				// _UNIT_TESTING
 	return sc_main(argc, argv);
-#endif				// _UNIT_TESTING
 }
