+++ src/utils_test.c
@@ -1,0 +1,2 @@
+#include "utils.h"
+#include "utils.c"
